[SDA-532] - update add level design page
